Rename all assessment components into survey
