+++ src/widgets/Survey/ui/SummaryScreen/Alerts.tsx
